requirements.txt
apturl==0.5.2
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-cors-headers==3.11.0
django-crontab==0.7.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
djoser==2.1.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
fuzzywuzzy==0.18.0
httplib2==0.14.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.1.0
protobuf==3.6.1
psycopg2-binary==2.9.3
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-Levenshtein==0.12.2
python3-openid==3.2.0
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.3
xkit==0.0.0
youtube-dl==2020.3.24
